no comment

+++ flutter_mvvm_simple_app/packages/app_theme/lib/src/extensions/color.dart
@@ -17,7 +17,7 @@ extension HexColor on Color {
   static Color shiftHsl(Color c, [double amt = 0]) {
     final hslc = HSLColor.fromColor(c);
     return hslc
-        .withLightness((hslc.lightness + amt).clamp(0.0, 1.0) as double)
+        .withLightness((hslc.lightness + amt).clamp(0.0, 1.0))
         .toColor();
   }
